fastapi-int8/app/requirements.txt
fastapi
uvicorn
requests
Pillow
numpy
opencv-python
onnx
onnxruntime
opencv-python